requirements.txt
adrf==0.1.3
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
crispy-bootstrap4==2023.1
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.3
distro==1.8.0
Django==5.0.1
django-crispy-forms==2.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
fake-useragent==1.4.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==7.0.1
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langcodes==3.3.0
langsmith==0.0.83
lxml==5.1.0
MarkupSafe==2.1.3
marshmallow==3.20.2
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.9.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parse==1.20.0
pinecone-client==3.0.2
preshed==3.0.9
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.2
tiktoken==0.5.2
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
wasabi==1.1.2
weasel==0.3.4
websockets==10.4
whitenoise==6.6.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
